+++ clemcore/playpen/__init__.py
@@ -1,6 +1,8 @@
 from clemcore.playpen.callbacks import BaseCallback, GameRecordCallback, RolloutProgressCallback, CallbackList
 from clemcore.playpen.base import BasePlayPen
-from clemcore.playpen.envs.env import GameEnv, make_env
+from clemcore.playpen.envs import make_env, PlayPenEnv
+from clemcore.playpen.envs.game_env import GameEnv
+from clemcore.playpen.envs.tree_env import GameTreeEnv
 
 __all__ = [
     "BaseCallback",
@@ -8,6 +10,8 @@ __all__ = [
     "RolloutProgressCallback",
     "CallbackList",
     "BasePlayPen",
+    "PlayPenEnv",
     "GameEnv",
+    "GameTreeEnv",
     "make_env"
 ]
