--- clemcore/playpen/envs/game_tree.py
@@ -1,69 +1,0 @@
-from typing import List, Dict
-
-from clemcore.playpen import GameEnv
-
-
-class GameTreeNode:
-    def __init__(self, game_env: GameEnv):
-        self._game_env = game_env
-        self._branches: List[GameTreeNode] = []
-        self._parent: GameTreeNode = self  # root is its own parent
-
-    def __iter__(self):
-        return iter(self._branches)
-
-    def __bool__(self):
-        return bool(self._branches)
-
-    def unwrap(self):
-        return self._game_env
-
-    def wraps(self, game_env: GameEnv) -> bool:
-        return self._game_env is game_env
-
-    def add_branch(self, branch_node: "GameTreeNode"):
-        self._branches.append(branch_node)
-        branch_node._parent = self
-
-
-class GameTree:
-
-    def __init__(self, root: GameEnv):
-        self._root: GameTreeNode = GameTreeNode(root)
-
-    def reset(self):
-        ...  # todo
-
-    def find_node(self, target_env: GameEnv):
-        def _find_node(node):
-            if node.wraps(target_env):  # check for object identity
-                return node
-
-            for branch in node:
-                target_node = _find_node(branch)
-                if target_node:
-                    return target_node
-
-            return None
-
-        return _find_node(self._root)
-
-    def find_leaves(self, unwrap=False):
-        def _find_leaves(node):
-            if not node:
-                if unwrap:
-                    return [node.unwrap()]
-                return [node]
-            leaves = []
-            for branch in node:
-                leaves.extend(_find_leaves(branch))
-            return leaves
-
-        return _find_leaves(self._root)
-
-    def add_branch(self, parent_env, branch_env):
-        # Find parent node and add child
-        parent_node = self.find_node(parent_env)
-        assert parent_node is not None, "There must be a parent node that wraps the candidates parent env"
-        branch_node = GameTreeNode(branch_env)
-        parent_node.add_branch(branch_node)
